new concept
I still have trouble to get the information out of the Info struct :S
