+++ samples/mpd/mpd.go
@@ -24,16 +24,32 @@ const (
 	PausedIcon  = PlaybackStatusIcon("⏸")
 )
 
+type MPD struct {
+	watcher            *mpd.Watcher
+	Message            chan *Info
+}
+
 type Info struct {
 	PlaybackStatus     PlaybackStatus
 	Title              string
 	Artist             string
 	PlaybackStatusIcon PlaybackStatusIcon
-	conn               *mpd.Client
-	watcher            *mpd.Watcher
 }
 
-func (i *Info) setupConnection(host string) {
+// Paused returns true if the media player is connected and paused.
+func (i Info) Paused() bool { return i.PlaybackStatus == Paused }
+
+// Playing returns true if the media player is connected and playing media.
+func (i Info) Playing() bool { return i.PlaybackStatus == Playing }
+
+// Stopped returns true if the media player is connected but stopped.
+func (i Info) Stopped() bool { return i.PlaybackStatus == Stopped }
+
+// Connected returns true if the media player is connected.
+func (i Info) Connected() bool { return i.PlaybackStatus != Disconnected }
+
+
+func connect(host string) *mpd.Client {
 	if host == "" {
 		host = "127.0.0.1:6600"
 	}
@@ -41,12 +57,36 @@ func (i *Info) setupConnection(host string) {
 	if err != nil {
 		conn = nil
 	}
-	i.conn = conn
+	return conn
+}
+
+func getStatus(host string) (i *Info){
+	conn := connect(host)
+	status, _ := conn.Status()
+	i.PlaybackStatus = PlaybackStatus(status["state"])
+	if i.Playing() {
+		i.PlaybackStatusIcon = PlayingIcon
+	} else if i.Paused() {
+		i.PlaybackStatusIcon = PausedIcon
+	}
+	if i.Playing() || i.Paused() {
+		currentSong, _ := conn.CurrentSong()
+		i.Artist = currentSong["Artist"]
+		i.Title = currentSong["Title"]
+	}
+	return i
+}
+
+func (m *MPD) watch(host string) {
 	watcher, err := mpd.NewWatcher("tcp", host, "")
 	if err != nil {
-		i.watcher = nil
+		m.watcher = nil
+		return
 	} else {
-		i.watcher = watcher
+		m.watcher = watcher
+	}
+	for {
+		m.Message <- getStatus(host)
 	}
 }
 
@@ -62,9 +102,11 @@ func (m *Module) Output(outputFunc func(Info) bar.Output) *Module {
 }
 
 func (m *Module) Stream(s bar.Sink) {
-	var i Info
-	i.setupConnection(m.host)
-	i.updateInfo()
+	mpdDaemon := &MPD{}
+	go func() {
+		mpdDaemon.watch(m.host)
+	}()
+	i := <-mpdDaemon.Message
 	tmp := outputs.Textf("%s %s - %s", i.PlaybackStatusIcon, i.Artist, i.Title)
 	s.Output(tmp)
 }
